requirements.txt
# PhotoSpot Korea - Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Scheduling
apscheduler>=3.10.4

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0

# Data Processing
pydantic>=2.5.0
python-dateutil>=2.8.2

# Gemini AI
google-generativeai>=0.3.0

# Telegram
python-telegram-bot>=20.7

# Utilities
python-dotenv>=1.0.0
orjson>=3.9.10

# Astronomy (for moon phase, sunrise/sunset)
astral>=3.2

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
